refactor(+page): call refresh api on layout instead of specific project page
