+++ src/mapper/src/routes/+layout.ts
@@ -1,14 +1,34 @@
 import type { PageLoad } from './$types';
-
+import { getLoginStore } from '$store/login.svelte.ts';
+import { refreshCookies, getUserDetailsFromApi } from '$lib/utils/login';
 // NOTE we can't prerender as we are using dynamic routing [projectId]
 export const prerender = false;
 export const ssr = false;
 
 import 'virtual:uno.css';
+import { error } from '@sveltejs/kit';
 // import { initDb } from '$lib/db/pglite';
 
 export const load: PageLoad = async ({ fetch }) => {
 	let config;
+	const loginStore = getLoginStore();
+
+	try {
+		/*
+			Login + user details
+		*/
+		let apiUser = await refreshCookies(fetch);
+		if (apiUser?.username !== 'svcfmtm') {
+			// Call /auth/me to populate the user details in the header
+			apiUser = await getUserDetailsFromApi(fetch);
+			if (!apiUser) {
+				loginStore.signOut();
+				throw error(401, { message: `You must log in first` });
+			} else {
+				loginStore.setAuthDetails(apiUser);
+			}
+		}
+	} catch (error) {}
 
 	try {
 		const s3Response = await fetch(`${import.meta.env.VITE_S3_URL}/fmtm-data/frontend/config.json`);
